refac(component): extract header as a component
