+++ ReactNativeStudy/MJ/Week1/todoApp/src/components/navigation/StackNavigation.tsx
@@ -1,7 +1,7 @@
 import {createNativeStackNavigator} from '@react-navigation/native-stack';
 import React from 'react';
 import {useRecoilValue} from 'recoil';
-import AddTodo from '../../screens/AddTodo';
+import AddTodo from '../../screens/AddTodo/AddTodo';
 import colorState from '../../store/color';
 import palette from '../../styles/palette';
 import {RootStackParamList} from '../../types/type';
